gcc 3.2.3 complains on long long unsigned in format, so type-cast __uint64_t to UINT64
git-svn-id: http://svn.open64.net/svnroot/open64@320 ead328b1-e903-0410-a19d-eacee4b031f6

+++ projects/pathscale-2.2.1/kpro64/be/cg/ia64/cgdwarf_targ.cxx
@@ -1587,7 +1587,7 @@ unwind_dump2asm (char *unwind_table_ptr,
 	// dump section in 8-byte chunks
 	fprintf(Asm_File, ".Lunwind_info_%d:\n", Current_PU_Count());
 	for (i = last_info_size; i < unwind_info_size; i+=8) {
-		fprintf(Asm_File, "\t%s %#llx\n", AS_DWORD, *(__uint64_t *)(unwind_info_ptr+i));
+		fprintf(Asm_File, "\t%s %#llx\n", AS_DWORD, (UINT64)*(__uint64_t *)(unwind_info_ptr+i));
 	}
 	fprintf(Asm_File, "\t%s %s\n", AS_SECTION, IA64_UNWIND);
 	// should always be 3 double-words
